requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlmodel==0.0.22
asyncpg
pydantic==2.10.1
pydantic-settings==2.7.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
sqlalchemy==2.0.36
cryptography==43.0.1
python-dotenv==1.0.1
PyJWT==2.8.0
psycopg[binary]==3.2.3
greenlet